Separate chunking and fix errors
