+++ src/components/Reminder/ReminderModal.vue
@@ -82,7 +82,7 @@ export default class ReminderModal extends Vue {
     } else {
       Vue.set(vxm.store.reminders, editedIndex, { ...this.reminder });
     }
-    vxm.store.updateChunks();
+    vxm.store.chunks.update();
     vxm.store.uploadReminders();
   }
 
@@ -94,7 +94,7 @@ export default class ReminderModal extends Vue {
 
     // Using Vue.delete ensures reactivity
     Vue.delete(vxm.store.reminders, this.editedIndex);
-    vxm.store.updateChunks(); // Update chunks
+    vxm.store.chunks.update(); // Update chunks
     vxm.store.uploadReminders();
   }
